backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
SQLAlchemy==2.0.45
starlette==0.50.0
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
gunicorn==21.2.0
APScheduler==3.10.4
